+++ ext/hallon/extconf.rb
@@ -1,15 +1,31 @@
 require 'mkmf'
 
+# Miscellaneous
 def error(message)
   abort "[ERROR] #{message}"
 end
 
 $CFLAGS << ' -ggdb -O0 -Wextra'
 
-error 'Missing ruby header' unless have_header 'ruby.h'
-error 'Missing OpenAL/alc.h' unless have_header 'OpenAL/alc.h'
-error 'Missing OpenAL/al.h' unless have_header 'OpenAL/al.h'
+# Check for headers
 
-with_ldflags('-framework OpenAL') { RUBY_PLATFORM =~ /darwin/ }
+error 'Missing ruby header' unless have_header 'ruby.h'
+
+if have_header('OpenAL/alc.h')
+  # yay, probably mac os!
+elsif have_header('AL/alc.h')
+  # woot, probably everybody else!
+else
+  error 'Missing openal headers'
+end
+
+# Add library
+if RUBY_PLATFORM =~ /darwin/
+  $LDFLAGS << ' -framework OpenAL '
+elsif have_library('openal', 'alSourceQueueBuffers')
+  $LDFLAGS << ' -lopenal '
+else
+  error 'Missing openal library'
+end
 
 create_makefile('openal_ext')