+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v0.0.3
+- Check for headers in multiple locations: OpenAL/ and AL/
+- Link against openal on non-OSX platforms
+
 # v0.0.2
 - Update dependency information; Hallon v0.12 or higher is required.